requirements.txt
###############################################################################
# Core Framework
###############################################################################
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-filter==25.1
django-cors-headers==4.7.0
django-csp==4.0
django-timezone-field==7.1
sqlparse==0.5.3
asgiref==3.9.1
tzdata==2025.2
Pillow==11.3.0

###############################################################################
# Django Extensions & Utilities
###############################################################################
django-auditlog==3.2.1
django-axes==8.0.0
django-celery-beat==2.8.1
django_celery_results==2.6.0
django-redis==6.0.0
django-health-check==3.20.0
django-otp==1.6.1
django-request-id==1.0.0
django-simple-history==3.10.1
django-storages==1.14.6
dj-database-url==3.0.1

###############################################################################
# Async, Task Queue & Scheduling (Celery + Redis)
###############################################################################
celery==5.5.3
billiard==4.2.1
kombu==5.5.4
vine==5.1.0
redis==6.4.0
cron-descriptor==1.4.5
python-crontab==3.3.0
flower==2.0.1

###############################################################################
# Server / Deployment
###############################################################################
gunicorn==23.0.0
tornado==6.5.2
gevent==25.5.1
greenlet==3.2.4

###############################################################################
# Security & Auth
###############################################################################
cryptography==45.0.6
PyJWT==2.10.1

###############################################################################
# API Documentation
###############################################################################
drf-spectacular==0.28.0
uritemplate==4.2.0

###############################################################################
# Cloud & Storage
###############################################################################
boto3==1.40.11
botocore==1.40.11
s3transfer==0.13.1
jmespath==1.0.1 

###############################################################################
# Configuration & Environment
###############################################################################
python-decouple==3.8
python-dotenv==1.1.1
PyYAML==6.0.2

###############################################################################
# Utilities
###############################################################################
attrs==25.3.0
idna==3.10
requests==2.32.4
charset-normalizer==3.4.3
urllib3==2.5.0
packaging==25.0
python-dateutil==2.9.0.post0
pytz==2025.2  
inflection==0.5.1
six==1.17.0
wcwidth==0.2.13
prompt_toolkit==3.0.51
typing_extensions==4.14.1 
humanize==4.12.3
prometheus_client==0.22.1  

###############################################################################
# Databases
###############################################################################
psycopg2==2.9.10
oracledb==3.3.0

###############################################################################
# Schema / Validation
###############################################################################
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
referencing==0.36.2
rpds-py==0.27.0

###############################################################################
# Monitoring / Observability
###############################################################################
sentry-sdk==2.35.0
psutil==7.0.0 
###############################################################################
# Serialization / Parsing
###############################################################################
cffi==1.17.1
pycparser==2.22

###############################################################################
# Messaging
###############################################################################
amqp==5.3.1

###############################################################################
# Click (CLI Utilities)
###############################################################################
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0

###############################################################################
# Zope (required by gevent / greenlet stack)
###############################################################################
zope.event==5.1.1
zope.interface==7.2

###############################################################################
# Build System Tools
###############################################################################
setuptools==80.9.0
pip==24.0
certifi==2025.8.3

###############################################################################
# Logging
##############################################################################
python-json-logger==3.3.0

###############################################################################
# Reporting & Data Processing
###############################################################################
reportlab==4.4.3
openpyxl==3.1.5
pandas==2.3.2
numpy==2.3.2 